requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.1
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
filter==0.0.0.20200724
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
Markdown==3.3.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.4.0
